+++ src/org/xjge/ui/Icon.java
@@ -213,7 +213,7 @@ public final class Icon {
         
         UIShader.getInstance().setUniform("uType", 3);
         UIShader.getInstance().setUniform("uOpacity", opacity);
-        UIShader.getInstance().setUniform("uColor", color.asVec3());
+        UIShader.getInstance().setUniform("uColor", color.asVector3f());
         UIShader.getInstance().setUniform("uModel", graphics.modelMatrix);
         UIShader.getInstance().setUniform("uProjection", UI.getProjectionMatrix());
         UIShader.getInstance().setUniform("uTexCoords", texCoords);